feature: implement test inButton component
