+++ apps/front/todo-nextjs/test/__tests__/component/container/Button/Button.test.tsx
@@ -1,7 +1,7 @@
-﻿import { cleanup, render, waitFor } from '@testing-library/react';
+﻿import { render, waitFor } from '@testing-library/react';
 import userEvent, { UserEvent } from '@testing-library/user-event';
 
-import ButtonContainer from '~/components/container/Button/Button';
+import Button from '~/components/container/Button/Button';
 
 const mockOnClick = jest.fn();
 
@@ -22,7 +22,7 @@ describe('~/component/container/Button.tsx', () => {
     mockOnClick.mockReset();
 
     const { getByRole } = render(
-      <ButtonContainer
+      <Button
         presentational={{
           text: mockProps.presentational.text,
           width: mockProps.presentational.width,
@@ -34,36 +34,6 @@ describe('~/component/container/Button.tsx', () => {
 
     user = userEvent.setup();
     button = getByRole('button');
-  });
-
-  it('propsで渡した文言でボタンを表示する。', async () => {
-    expect(button).toHaveTextContent('test');
-  });
-
-  it('propsで渡したサイズでボタンを表示する。', async () => {
-    expect(button).toHaveStyle({ width: '128px' });
-  });
-
-  it('活性化したボタンを表示する。', async () => {
-    expect(button).toBeEnabled();
-  });
-
-  it('非活性化したボタンを表示する。', async () => {
-    cleanup();
-
-    const { getByRole: getByRoleDisabled } = render(
-      <ButtonContainer
-        presentational={{
-          text: mockProps.presentational.text,
-          width: mockProps.presentational.width,
-        }}
-        disabled={true}
-        onClick={mockProps.onClick}
-      />,
-    );
-    const disabledButton = getByRoleDisabled('button');
-
-    expect(disabledButton).toBeDisabled();
   });
 
   it('ボタンを押下すると、propsで渡したonClickを発火する。', async () => {
